requirements.txt
python-telegram-bot
python-dotenv
gigachain
chromadb
bs4
fake_useragent
nest_asyncio
Spire.Doc
pypdf
